Android scan record parsing: fix length and type extraction
The scanRecordBuffer is an array of chars. In the scan record format,
the length of the entry is represented by one byte. On the platforms
where type 'char' is signed, having the value with the upper bit set
(greater than 127) will result in a negative length when extracted.
That could lead to the 'scanRecordLength - nBytes' check provide
incorrect results. Also, the code that modifies the i variable could
behave wrong.

The similar problem exists for the advertising type byte. Types with
the upper bit set (e.g. manufacturer data) can be interpreted
incorrectly.

Fix it by explicitly converting to quint8. Conveniently, ARM platforms
(typical for Android) treat 'char' as unsigned, but there is a chance
that these issue could appear on Android x86.

Extend the tests with the case that will loop infinitely if the bug
is not fixed.

Task-number: QTBUG-145934
Pick-to: 6.11 6.8 6.5 5.15

+++ src/bluetooth/android/devicediscoverybroadcastreceiver.cpp
@@ -430,14 +430,14 @@ Q_BLUETOOTH_EXPORT void parseScanRecord(QBluetoothDeviceInfo &info, jbyteArray s
         // sizeof(EIR Data) = sizeof(Length) + sizeof(EIR data Type) + sizeof(EIR Data)
         // Length = sizeof(EIR data Type) + sizeof(EIR Data)
 
-        const int nBytes = scanRecordBuffer[i];
+        const int nBytes = quint8(scanRecordBuffer[i]);
         if (nBytes == 0)
             break;
 
         if (i >= scanRecordLength - nBytes)
             break;
 
-        const int adType = scanRecordBuffer[i+1];
+        const int adType = quint8(scanRecordBuffer[i+1]);
         const char *dataPtr = &scanRecordBuffer[i+2];
         QBluetoothUuid foundService;
